test: unit tests resolved
